refactor: 총 수입/지출 컴포넌트 props 방식 변경
